api changes.

+++ IPv6Addressing/lib/vmem/rmem.h
@@ -31,7 +31,7 @@ Write exactly BLOCK_SIZE bytes to a given block on the virtual disk.
 and "data" is a pointer to the data to write.
 */
 
-void rmem_write( struct rmem *r, int block, char *data );
+void rmem_write( struct rmem *r, uint64_t block, char *data );
 
 /*
 Read exactly BLOCK_SIZE bytes from a given block on the virtual disk.
@@ -39,13 +39,13 @@ Read exactly BLOCK_SIZE bytes from a given block on the virtual disk.
 and "data" is a pointer to where the data will be placed.
 */
 
-void rmem_read( struct rmem *r, int block, char *data );
+void rmem_read( struct rmem *r, uint64_t block, char *data );
 
 /*
 Return the number of blocks in the virtual disk.
 */
 
-int rmem_blocks( struct rmem *r );
+uint64_t rmem_blocks( struct rmem *r );
 
 /*
 Close the virtual disk.
